chore: port more tests to jest
