+++ jest-tests/addMany-dynamic-template-file/plop-templates/bar-chart/helpers/not-included.txt
@@ -1,0 +1,1 @@
+This file should not be created.
