+++ jest-tests/addMany-dynamic-template-file/plop-templates/bar-chart/{{dashCase name}}-bar-tmpl.html
@@ -1,0 +1,2 @@
+this is a bar chart
+name: {{name}}
