+++ jest-tests/addMany-dynamic-template-file/plop-templates/line-chart/{{dashCase name}}-line-tmpl.html
@@ -1,0 +1,2 @@
+this is a line chart
+name: {{name}}
