+++ jest-tests/addMany-dynamic-template-file/plopfile.js
