+++ jest-tests/addMany-executable-file/plop-templates/{{dashCase executableName}}.sh
