+++ jest-tests/addMany-executable-file/plopfile.js
