+++ jest-tests/addMany-multiple-files/plop-templates/add.txt
@@ -1,0 +1,1 @@
+name: {{titleCase name}}
