+++ jest-tests/addMany-multiple-files/plop-templates/components/logic/{{dashCase name}}-plop-logo.png
