+++ jest-tests/addMany-multiple-files/plop-templates/nested-folder/a-nested-add.txt
@@ -1,0 +1,1 @@
+constant name: {{constantCase name}}
