+++ jest-tests/addMany-non-verbose/plop-templates/add.txt
@@ -1,0 +1,1 @@
+name: {{titleCase name}}
