+++ jest-tests/addMany-non-verbose/plop-templates/components/logic/{{dashCase name}}-plop-logo.png
