+++ jest-tests/addMany-strip-extensions/plopfile.js
