+++ jest-tests/append/plop-templates/list.txt
@@ -1,0 +1,7 @@
+Don't remove me: Plop
+
+-- APPEND ITEMS HERE --
+
+/* APPEND OTHER ITEMS HERE */
+
++++++++++++++++++++++++++++++++++++++++
