--- tests/addMany-dynamic-template-file-mock/plop-templates/bar-chart/helpers/not-included.txt
@@ -1,1 +1,0 @@
-This file should not be created.
