--- tests/addMany-dynamic-template-file-mock/plop-templates/bar-chart/{{dashCase name}}-bar-tmpl.html
@@ -1,2 +1,0 @@
-this is a bar chart
-name: {{name}}
