--- tests/addMany-dynamic-template-file-mock/plop-templates/line-chart/{{dashCase name}}-line-tmpl.html
@@ -1,2 +1,0 @@
-this is a line chart
-name: {{name}}
