--- tests/addMany-multiple-files-mock/plop-templates/add.txt
@@ -1,1 +1,0 @@
-name: {{titleCase name}}
