--- tests/addMany-multiple-files-mock/plop-templates/nested-folder/a-nested-add.txt
@@ -1,1 +1,0 @@
-constant name: {{constantCase name}}
