--- tests/append-mock/plop-templates/list.txt
@@ -1,7 +1,0 @@
-Don't remove me: Plop
-
--- APPEND ITEMS HERE --
-
-/* APPEND OTHER ITEMS HERE */
-
-+++++++++++++++++++++++++++++++++++++++
